Using MLP to solve MNIST digit dataset
In this program I coded a Machine Learning Neural Network from scracth. I applied a MLP brain that has customizable parameters. Can change modes that the program runs in the main() function. the draw() function predicts the user's drawn digits
Note: This is not the dataset i trained the model on, it is made smaller so I could upload it here