Recognize doodles
All the data was removed so that I could upload it to GitHub. So training the model and testing it against test data is not possible but the code for everything is still there.
The only functionallity this program has is to let it predict your own drawings. It has a set of objects it can recognize so drawing anything besides those objects won't do anything.
This still uses an MLP only